reworked pari payouts into a totalizator with coefficients
Фиксированный множитель x2 заменен тотализатором из calc.md: ставки идут
в общий пул, из него удерживается комиссия организатора, а призовой фонд
делится между победителями пропорционально их ставкам.

- PariPayoutCalculator: комиссия, призовой фонд, коэффициент и выплата;
  целочисленная арифметика с округлением вниз — сумма выплат не превышает
  призовой фонд
- итоги розыгрыша (total_pool, prize_pool, winning_sum, коэффициент)
  фиксируются один раз в finish() под блокировкой пари; расчет по ставке
  опирается только на них, поэтому остается идемпотентным
- если на победивший вариант никто не поставил — возврат всем без комиссии
- commission_rate фиксируется в пари при создании (PARI_COMMISSION_RATE,
  по умолчанию 5%), изменение настройки не влияет на идущие пари
- в опросе показываются живые коэффициенты и удержанная комиссия,
  в футере завершенного пари — итоговый коэффициент
- V5: колонки итогов, commission_rate, points_transactions.amount -> BIGINT
  (выплата больше не ограничена удвоенной ставкой), бэкфилл итогов для уже
  завершенных пари
- 200 тестов, 95% инструкций / 91% ветвей; миграция V5 проверена накатом
  на базу со схемой V4 и данными

+++ src/main/java/ru/z3r0ing/discordlp/command/LpKickCommand.java
@@ -71,7 +71,7 @@ public class LpKickCommand implements SlashCommandHandler {
 
         PointsTransaction tx = new PointsTransaction();
         tx.setMember(callerMember);
-        tx.setAmount(-KICK_COST);
+        tx.setAmount((long) -KICK_COST);
         tx.setReason(TransactionReason.USER_KICK);
         tx.setCreatedAt(Instant.now());
         pointsTransactionRepository.save(tx);
